// src/lib/api/generated/model/createMilestoneDtoStage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * BEACON Work Progress API
 * Secure client dashboard API - projects, milestones, updates, feedback.
 * OpenAPI spec version: 1.0
 */

export type CreateMilestoneDtoStage = typeof CreateMilestoneDtoStage[keyof typeof CreateMilestoneDtoStage];


export const CreateMilestoneDtoStage = {
  DISCOVER: 'DISCOVER',
  DESIGN: 'DESIGN',
  BUILD: 'BUILD',
  SCALE: 'SCALE',
} as const;