// src/lib/api/generated/model/feedbackResponseDtoStatus.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * BEACON Work Progress API
 * Secure client dashboard API - projects, milestones, updates, feedback.
 * OpenAPI spec version: 1.0
 */

export type FeedbackResponseDtoStatus = typeof FeedbackResponseDtoStatus[keyof typeof FeedbackResponseDtoStatus];


export const FeedbackResponseDtoStatus = {
  OPEN: 'OPEN',
  ADDRESSED: 'ADDRESSED',
} as const;